// src/hooks/kernel/keyboard/enums.ts
export enum KeyboardEvent {
    INPUT = 'input',
    DELETE = 'delete',
    UNDO = 'undo',
    SPACE = 'space',
    COMBO = 'combo',
    MOVE = 'move',
    ENTER = 'enter',
    NUMBER = 'number',
    FUNCTION = 'fn',
    PAGE = 'page',
    HOME = 'home',
    CLEAR = 'clear',
    OPERATOR = 'operator',
    CUSTOM = 'custom',
    CUSTOM_ALL = 'custom_all'
}